services/syslog_trap_receiver/requirements.txt
psycopg2-binary==2.9.9
asyncio==3.4.3
pysnmp==4.4.12
pysmi==0.3.4
